Diff the CloudFormation stacks on PR
Inspired by https://github.com/douglasnaphas/madliberation/issues/343.

This prints a `cdk diff` in the test account and prod account on PRs.

That way, when someone raises a PR, you can see what changes will be
made to your AWS resources in the test and prod account if the PR is
merged.

Similarly, if you want to see how a change will affect your resources,
you can raise a PR with the change, and look at the GitHub Actions
output for the PR build.

Note that we don't want to deploy any resources in PR builds, because
then someone could come along and create some expensive resource in a
PR, and it would get created in an account under our control.

Here is an example of a PR build with `cdk diff` in action:

https://github.com/douglasnaphas/madliberation/pull/344/checks?check_run_id=2903164159